refactor: extract ProfileFetcher into AppState extension (Stage 3b)
Stage 3b of the AppState god-object split. Largest of the three Stage 3
sub-extractions — moves all kind:0 profile metadata + image cache +
multi-relay fetch + cross-relay merge logic out of AppState.swift into
a dedicated extension file.

Moves out of Clave/AppState.swift:
- loadCachedProfileImage() — disk image cache loader
- updateCurrentProfile(_:) — write-through to accounts list (private)
- cachedImageURL(for:) func + cachedImageURL var (computed)
- cacheImage(from:pubkey:) — Bug F invariant preserved (explicit pubkey)
- fetchProfile(for:) instance — multi-relay parallel fanout
- fetchProfileIfNeeded() — 1h throttled current-account fetch
- refreshProfile(for:) + refreshProfileAsync(for:) — on-demand refresh
- fetchProfilesForAllAccountsIfNeeded() — all-accounts throttled
- refreshAllProfiles() — all-accounts pull-to-refresh
- struct FetchedKind0 — nested type used by tests
- static mergeKind0(_:) — pure cross-relay selection (already unit-tested)
- static fetchProfile(from:pubkey:) — single-relay kind:0 fetch

Into the new file:
- Clave/AppState+ProfileFetcher.swift (303 LOC) — extension AppState

What stays in AppState.swift:
- var profile (computed) — kept on the class to preserve @Observable
  macro behavior intact (computed properties in extensions on @Observable
  classes is an untested edge; defensive 3-line cost). Pointer comment
  added.
- var profileImage (stored) — Swift forbids stored properties in
  extensions. Pointer comment added.
- The 4 reset call sites for `profileImage = nil` (in account
  switching/deleting flows) — they remain in AppState.swift and resolve
  through the @Observable property setter.

Also dropped `private` modifier on `persistAccounts()` — required because
the extension's `updateCurrentProfile` and `fetchProfile(for:)` call it
across files. `private` doesn't cross file boundaries even within
extensions of the same type. Will move to AccountManager in Stage 4.

Zero behavior change:
- Function bodies preserved byte-for-byte.
- Original `private` modifiers preserved on file-local methods
  (updateCurrentProfile, cachedImageURL var, cacheImage, fetchProfile
  instance + static).
- @MainActor on refreshProfileAsync preserved.
- All Shared/ dependencies (LightRelay) and AppState calls
  (persistAccounts, accounts, currentAccount, profileImage) work
  unchanged from the extension.

Test impact: zero changes.
- AppStateProfileMergeTests references AppState.FetchedKind0 +
  AppState.mergeKind0 — both still resolve correctly with the
  nested-type-in-extension pattern.

External callers unchanged:
- HomeView.swift:120,124 (refreshAllProfiles, fetchProfilesForAllAccountsIfNeeded)
- AccountDetailView.swift:65 (refreshProfileAsync)
- MultiAccountDiagnosticsView.swift:154 (fetchProfileIfNeeded)
All call via appState.<method> — works as extension method dispatch.

No pbxproj edits — Clave/ directory is auto-synced.

Verification:
- xcodebuild test -skip-testing:ClaveUITests on iPhone 17 / iOS 26.4:
  - Pre-baseline (main @ d8b5a55): 232 passed / 0 failed
  - Post-extraction (this branch): 232 passed / 0 failed
  - ** TEST SUCCEEDED ** in both runs

AppState.swift: 1,770 -> 1,483 LOC (-287, -16.2%).
Combined Stages 1+2+3a+3b: 2,338 -> 1,483 LOC (-36.6% from original).

Stage 3c (ProxyClient, ~400 LOC) follows.
